+++ mobile/android/android-components/components/feature/accounts/src/main/java/mozilla/components/feature/accounts/FirefoxAccountsAuthFeature.kt
@@ -65,6 +65,7 @@ class FirefoxAccountsAuthFeature(
         override fun onLoadRequest(
             engineSession: EngineSession,
             uri: String,
+            lastUri: String?,
             hasUserGesture: Boolean,
             isSameDomain: Boolean,
             isRedirect: Boolean,